Revert "More styling, changed admin toggle due to bugs"
