+++ routes/Cart/index.js
@@ -6,19 +6,13 @@ const {
   createCartItem,
   deleteCartItem,
   updateProductQuantity,
-  getUserById,
 } = require("../../db");
-
-cartRouter.get("/", requireUser, async (req, res, next) => {
-  const { id } = req.user;
-  const user = await getUserById(id);
-  res.send(user.cart);
-});
 
 cartRouter.post("/", requireUser, async (req, res, next) => {
   const { id } = req.user;
   const { product_id, quantity } = req.body;
   const addedItem = await addProductToCart(id, product_id, quantity);
+  console.log(addedItem);
   res.send(addedItem);
 });
 
@@ -33,7 +27,6 @@ cartRouter.patch("/:productId", requireUser, async (req, res, next) => {
   const { id } = req.user;
   const { productId } = req.params;
   const { quantity } = req.body;
-  console.log(id, productId, quantity);
   const updatedItem = await updateProductQuantity(id, productId, quantity);
   res.send(updatedItem);
 });
