+++ routes/Users/index.js
@@ -10,6 +10,7 @@ const {
   createUser,
   getUserByUsername,
   verifyUniqueUser,
+  client,
 } = require("../../db");
 const { patchUser } = require("../../db/users/index.js");
 
@@ -68,18 +69,26 @@ usersRouter.post("/register", async (req, res, next) => {
   const { username, password, email, name } = req.body;
 
   try {
-    const existingUser = await verifyUniqueUser(username, email);
+    const existingUser = await verifyUniqueUser(username, email, name);
 
     if (existingUser != undefined) {
+      console.log("User exists");
       if (existingUser.username === username) {
         next({
           name: "DuplicateUsername",
           message: "This username is already in use",
         });
-      } else if (existingUser.email === email) {
+      }
+      if (existingUser.email === email) {
         next({
           name: "DuplicateEmail",
           message: "This email is already in use!",
+        });
+      }
+      if (existingUser.name === name) {
+        next({
+          name: "DuplicateName",
+          message: "Someone already has your name. Get a new one",
         });
       }
     } else {
@@ -116,7 +125,6 @@ usersRouter.patch("/:id", async (req, res, next) => {
     const fields = {
       admin,
     };
-    console.log(admin);
     const updatedUser = await patchUser(id, fields);
     res.send(updatedUser);
   } catch (error) {
