+++ src/App.js
@@ -1,7 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { BrowserRouter as Router, Link, Switch, Route } from "react-router-dom";
 import {
-  Cart,
   Product,
   Users,
   Login,
@@ -18,7 +17,6 @@ import "./app.css";
 const App = () => {
   const [loggedIn, setLoggedIn] = useState(false);
   const [admin, setAdmin] = useState(false);
-  const [cart, setCart] = useState([]);
 
   useEffect(() => {
     if (localStorage.getItem("token")) {
@@ -88,10 +86,7 @@ const App = () => {
         <main>
           <Switch>
             <Route exact path="/">
-              <Product cart={cart} setCart={setCart} />
-            </Route>
-            <Route exact path="/cart">
-              <Cart cart={cart} setCart={setCart} loggedIn={loggedIn} />
+              <Product />
             </Route>
             <Route exact path="/keyboard">
               <Keyboards />
