+++ src/Img/index.js
@@ -1,3 +1,2 @@
 export { default as Egg } from "./egg.png";
 export { default as AddtoCart } from "./addToCart.svg";
-export { default as TrashCan } from "./trash.svg";
