+++ src/api/index.js
@@ -56,6 +56,7 @@ export async function userRegister(name, email, username, password) {
       username,
       password,
     });
+    console.log(data.token);
     if (data.token) {
       alert("You have successfully registered!");
       setToken(data.token);
@@ -68,14 +69,15 @@ export async function userRegister(name, email, username, password) {
 
 export async function changeAdmin(id, admin) {
   try {
-    let updatedInfo = {
-      admin,
-    };
-    // let adminStatus = prompt(
-    //   "What would you like to change the admin status to?",
-    //   admin
-    // );
+    let updatedInfo = {};
+    let adminStatus = prompt(
+      "What would you like to change the admin status to?",
+      admin
+    );
 
+    if (adminStatus) {
+      updatedInfo.admin = adminStatus;
+    }
     const { data } = await axios.patch(`/api/users/${id}`, updatedInfo);
     console.log(data);
     return data;
@@ -97,61 +99,5 @@ export async function createProduct(name, description, price, image_url, type) {
     return data;
   } catch (error) {
     throw error;
-  }
-}
-
-export async function getCart(token) {
-  try {
-    const { data } = await axios.get("/api/cart", {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    });
-    console.log(data);
-    return data;
-  } catch (error) {
-    console.error("Error getting cart");
-  }
-}
-
-export async function updateProductQuantity(product_id, quantity, token) {
-  try {
-    const { data } = await axios.patch(
-      `/api/cart/${product_id}`,
-      { quantity },
-      {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      }
-    );
-    return data;
-  } catch (error) {
-    console.error("Error updating quantity");
-  }
-}
-
-export async function removeItemFromCart(product_id, token) {
-  try {
-    const { data } = await axios.delete(`api/cart/${product_id}`, {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    });
-    return data;
-  } catch (error) {
-    console.error("Error removing from cart");
-  }
-}
-
-export async function addItemToCart(product_id, quantity, token) {
-  try {
-    const { data } = await axios.post(
-      `api/cart`,
-      { product_id, quantity },
-      { headers: { Authorization: `Bearer ${token}` } }
-    );
-  } catch (error) {
-    console.error("Error adding to cart");
   }
 }
