--- src/components/Cart/Cart.css
@@ -1,86 +1,0 @@
-.CartCards {
-  display: flex;
-  flex-direction: column;
-}
-
-.CartCard {
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  background: rgb(255, 255, 255);
-  border-bottom: 1px solid #eee;
-}
-
-.quantity {
-  font-size: 20px;
-}
-
-.Checkout-Container {
-  display: flex;
-  flex-direction: row-reverse;
-  justify-content: flex-start;
-  align-items: center;
-
-  background-color: #3f3f3f;
-  color: #ebebeb;
-  /* width: 100vw;
-  bottom: 0;
-position: fixed; */
-}
-
-.QuantityContainer {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  min-width: 10%;
-}
-
-.QuantityInput {
-  margin: 20px;
-  text-align: center;
-  font-size: 24px;
-  border: 1px solid #eee;
-  border-radius: 10px;
-  height: 50px;
-  width: 50px;
-  background-color: rgb(243, 236, 236);
-}
-input::-webkit-outer-spin-button,
-input::-webkit-inner-spin-button {
-  -webkit-appearance: none;
-  margin: 0;
-}
-
-/* Firefox */
-input[type="number"] {
-  -moz-appearance: textfield;
-}
-.TotalPrice {
-  padding: 20px;
-  margin-right: 30px;
-}
-
-.CheckoutBtn {
-  height: 50px;
-  width: 200px;
-  background-color: #ebebeb;
-  color: #3f3f3f;
-  border-radius: 10px;
-  font-size: 20px;
-  transition: 0.1s ease;
-}
-
-.CheckoutBtn:hover {
-  background-color: white;
-  transform: scale(1.1);
-}
-
-.deleteIcon {
-  height: 50px;
-  width: 50px;
-}
-
-.deleteIcon:hover {
-  transform: scale(1.2);
-}
