+++ src/components/Login.js
@@ -1,33 +1,16 @@
 import React, { useState } from "react";
 import { Redirect } from "react-router-dom";
-import { userRegister } from "../../api";
-import "./Register.css";
+import { userLogin } from "../api";
 
-const Register = (props) => {
-  const { setLoggedIn } = props;
+const Login = (props) => {
+  const { setLoggedIn, setAdmin, message, setMessage } = props;
 
   const [username, setUsername] = useState("");
   const [password, setPassword] = useState("");
-  const [name, setName] = useState("");
-  const [email, setEmail] = useState("");
 
   return (
-    <form className="register" onSubmit={(event) => event.preventDefault()}>
-      <h1>Register:</h1>
-      <label>Name:</label>
-      <input
-        type="text"
-        value={name}
-        onChange={(event) => setName(event.target.value)}
-        placeholder=""
-      />
-      <label>Email:</label>
-      <input
-        type="text"
-        value={email}
-        onChange={(event) => setEmail(event.target.value)}
-        placeholder=""
-      />
+    <form className="login" onSubmit={(event) => event.preventDefault()}>
+      <h1>Login:</h1>
       <label>Username:</label>
       <input
         type="text"
@@ -46,9 +29,11 @@ const Register = (props) => {
         onClick={async (event) => {
           event.preventDefault();
           try {
-            let submit = await userRegister(name, email, username, password);
-            if (submit.message) {
-              alert(submit.message);
+            let submit = await userLogin(username, password);
+            setAdmin(submit.user.admin);
+
+            if (submit.error) {
+              alert("There was an error logging in...");
             } else {
               setLoggedIn(true);
               return <Redirect to="/" />;
@@ -58,10 +43,10 @@ const Register = (props) => {
           }
         }}
       >
-        Register
+        Login
       </button>
     </form>
   );
 };
 
-export default Register;
+export default Login;
